+++ ui/hooks/bridge/useLatestBalance.ts
@@ -1,5 +1,5 @@
 import { type Hex, type CaipChainId } from '@metamask/utils';
-import { useMemo } from 'react';
+import { useCallback, useMemo } from 'react';
 import {
   isSolanaChainId,
   calcLatestSrcBalance,
@@ -8,6 +8,7 @@ import {
   ChainId,
   isNativeAddress,
 } from '@metamask/bridge-controller';
+import { useLocation } from 'react-router-dom';
 import { getSelectedInternalAccount } from '../../selectors';
 import { useAsyncResult } from '../useAsync';
 import { Numeric } from '../../../shared/modules/Numeric';
@@ -31,10 +32,18 @@ const useLatestBalance = (
     decimals: number;
     symbol: string;
     string?: string;
-    chainId?: Hex | CaipChainId | ChainId;
+    chainId: Hex | CaipChainId | ChainId;
     assetId?: string;
   } | null,
 ) => {
+  const { search } = useLocation();
+
+  const searchParams = useMemo(() => new URLSearchParams(search), [search]);
+  const tokenAddressFromUrl = useMemo(
+    () => searchParams.get('token'),
+    [searchParams],
+  );
+
   const { address: selectedAddress, id } = useMultichainSelector(
     getSelectedInternalAccount,
   );
@@ -44,27 +53,26 @@ const useLatestBalance = (
     getMultichainBalances,
   );
 
-  const nonEvmBalances = nonEvmBalancesByAccountId?.[id];
+  const nonEvmBalances = useMemo(
+    () => nonEvmBalancesByAccountId?.[id],
+    [nonEvmBalancesByAccountId, id],
+  );
+
+  // Don't fetch the balance the src token balance until the tokenAddressFromUrl is unset
+  const shouldUpdateBalance = useCallback(
+    () => token?.chainId && !tokenAddressFromUrl,
+    [tokenAddressFromUrl, token],
+  );
 
   const value = useAsyncResult<string | undefined>(async () => {
-    if (!token?.chainId || !token) {
+    if (!shouldUpdateBalance() || !token) {
       return undefined;
     }
 
     const { chainId } = token;
-
-    // No need to fetch the balance for non-EVM tokens, use the balance provided by the
-    // multichain balances controller
-    if (isSolanaChainId(chainId) && token.decimals) {
-      const caipAssetType = isNativeAddress(token.address)
-        ? MULTICHAIN_NATIVE_CURRENCY_TO_CAIP19.SOL
-        : token.assetId ?? token.address;
-      return Numeric.from(
-        nonEvmBalances?.[caipAssetType]?.amount ?? token?.string,
-        10,
-      )
-        .shiftedBy(-1 * token.decimals)
-        .toString();
+    // No need to fetch the balance for non-EVM tokens
+    if (isSolanaChainId(chainId)) {
+      return undefined;
     }
 
     if (
@@ -82,7 +90,35 @@ const useLatestBalance = (
     }
 
     return undefined;
-  }, [currentChainId, token, selectedAddress, nonEvmBalances]);
+  }, [
+    currentChainId,
+    token,
+    selectedAddress,
+    tokenAddressFromUrl,
+    shouldUpdateBalance,
+  ]);
+
+  const nonEvmBalance = useMemo(() => {
+    if (!shouldUpdateBalance() || !token) {
+      return undefined;
+    }
+
+    const { chainId, decimals, address } = token;
+
+    // Use the balance provided by the multichain balances controller
+    if (isSolanaChainId(chainId) && decimals) {
+      const caipAssetType = isNativeAddress(address)
+        ? MULTICHAIN_NATIVE_CURRENCY_TO_CAIP19.SOL
+        : token.assetId ?? token.address;
+      return Numeric.from(
+        nonEvmBalances?.[caipAssetType]?.amount ?? token?.string,
+        10,
+      )
+        .shiftedBy(-1 * token.decimals)
+        .toString();
+    }
+    return undefined;
+  }, [shouldUpdateBalance, token, nonEvmBalances]);
 
   if (token && typeof token.decimals !== 'number') {
     throw new Error(
@@ -90,10 +126,15 @@ const useLatestBalance = (
     );
   }
 
+  const balance = useMemo(() => {
+    return token?.chainId && isSolanaChainId(token.chainId)
+      ? nonEvmBalance
+      : value?.value;
+  }, [nonEvmBalance, value?.value, token?.chainId]);
+
   return useMemo(
-    () =>
-      value?.value ? calcTokenAmount(value.value, token?.decimals) : undefined,
-    [value?.value, token?.decimals],
+    () => (balance ? calcTokenAmount(balance, token?.decimals) : undefined),
+    [balance, token?.decimals],
   );
 };